google_meet_bot/requirements.txt
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.1
PyAudio==0.2.14
pydantic==2.9.2
pydantic_core==2.23.4
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0
